Use size_t for index in util.cpp (int64_t for torch).

+++ src/util.cpp
@@ -26,9 +26,9 @@
 namespace rcgpar {
 std::vector<double> mixture_components(const seamat::Matrix<double> &probs, const std::vector<double> &log_times_observed) {
   std::vector<double> thetas(probs.get_rows(), 0.0);
-  for (uint32_t i = 0; i < probs.get_rows(); ++i) {
-    uint32_t n_times_total = 0.0;
-    for (uint32_t j = 0; j < probs.get_cols(); ++j) {
+  for (size_t i = 0; i < probs.get_rows(); ++i) {
+    size_t n_times_total = 0.0;
+    for (size_t j = 0; j < probs.get_cols(); ++j) {
       n_times_total += std::exp(log_times_observed[j]);
       thetas[i] += std::exp(probs(i, j) + log_times_observed[j]);
     }
@@ -44,8 +44,8 @@ std::vector<double> mixture_components_torch(const seamat::Matrix<double> &probs
   auto device = torch::cuda::is_available() ? torch::kCUDA : torch::kCPU;
   torch::Dtype precision = torch::kFloat64;
 
-  uint32_t rows = probs.get_rows();
-  uint32_t cols = probs.get_cols();
+  int64_t rows = probs.get_rows();
+  int64_t cols = probs.get_cols();
   
   torch::Tensor probs_ten = torch::from_blob((double*)probs_vec.data(), {rows, cols}, precision).clone().to(device);
   torch::Tensor log_times_observed_ten = torch::from_blob((double*)log_times_observed.data(), {cols}, precision).clone().to(device);
